RIX/main-agent/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.12.1
pgvector==0.2.4
# Scientific computing libraries - versions compatible with Python 3.9-3.11
numpy>=1.24.0,<2.1.0  # Python 3.9-3.11 compatible (2.1.0+ drops Python 3.9)
scikit-learn>=1.5.0,<1.7.0  # Python 3.9-3.13 compatible (1.7.0+ drops Python 3.9)
scipy>=1.11.0,<1.16.0  # Python 3.9-3.12 compatible (1.16.0+ drops Python 3.9/3.10)
threadpoolctl>=3.1.0,<4.0.0  # Threading control for scikit-learn

# HTTP Client
httpx>=0.25.0,<0.28.0
aiohttp>=3.9.1,<3.11.0

# WebSocket Support
websockets==12.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# NLP Libraries (pattern matching only)
nltk==3.8.1

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<0.24.0
black>=23.9.0,<25.0.0
flake8>=6.0.0,<8.0.0
mypy>=1.5.0,<2.0.0

# Production
gunicorn==21.2.0